requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.0.2
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-dajax==0.9.2
django-dajaxice==0.7
django-js-asset==1.2.2
idna==2.7
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
requests==2.20.1
urllib3==1.24.1
